migrated existing indexing tests to new class
